Make private things that can, remove some dead code.
