+++ creusot/src/translation.rs
@@ -1,10 +1,9 @@
 pub(crate) mod constant;
 pub(crate) mod external;
-#[allow(dead_code)]
 pub(crate) mod fmir;
 pub(crate) mod function;
 pub(crate) mod pearlite;
-pub(crate) mod projection_vec;
+mod projection_vec;
 pub(crate) mod specification;
 pub(crate) mod traits;
