+++ creusot/src/translation/traits.rs
@@ -1,22 +1,22 @@
 use super::pearlite::{Term, TermKind};
 use crate::{
     ctx::*,
-    translation::function::terminator::evaluate_additional_predicates,
     util::{is_law, is_spec},
 };
 use rustc_hir::def_id::DefId;
 use rustc_infer::{
     infer::{DefineOpaqueTypes, InferCtxt, TyCtxtInferExt},
-    traits::ObligationCause,
+    traits::{Obligation, ObligationCause, TraitEngine},
 };
 use rustc_middle::ty::{
     AssocItem, AssocItemContainer, Const, ConstKind, EarlyBinder, GenericArgs, GenericArgsRef,
-    ParamConst, ParamEnv, ParamTy, TraitRef, Ty, TyCtxt, TyKind, TypeFoldable, TypeFolder,
+    ParamConst, ParamEnv, ParamTy, Predicate, TraitRef, Ty, TyCtxt, TyKind, TypeFoldable,
+    TypeFolder,
 };
-use rustc_span::{Symbol, DUMMY_SP};
+use rustc_span::{Span, Symbol, DUMMY_SP};
 use rustc_trait_selection::{
     error_reporting::InferCtxtErrorExt,
-    traits::{orphan_check_trait_ref, ImplSource, InCrate},
+    traits::{orphan_check_trait_ref, FulfillmentError, ImplSource, InCrate, TraitEngineExt},
 };
 use rustc_type_ir::fold::TypeSuperFoldable;
 use std::collections::HashMap;
@@ -113,6 +113,28 @@ impl<'tcx> TranslationCtx<'tcx> {
         }
 
         TraitImpl { laws, refinements }
+    }
+}
+
+pub(crate) fn evaluate_additional_predicates<'tcx>(
+    infcx: &InferCtxt<'tcx>,
+    p: Vec<Predicate<'tcx>>,
+    param_env: ParamEnv<'tcx>,
+    sp: Span,
+) -> Result<(), Vec<FulfillmentError<'tcx>>> {
+    let mut fulfill_cx = <dyn TraitEngine<'tcx, _>>::new(infcx);
+    for predicate in p {
+        let predicate = infcx.tcx.erase_regions(predicate);
+        let cause = ObligationCause::dummy_with_span(sp);
+        let obligation = Obligation { cause, param_env, recursion_depth: 0, predicate };
+        // holds &= infcx.predicate_may_hold(&obligation);
+        fulfill_cx.register_predicate_obligation(&infcx, obligation);
+    }
+    let errors = fulfill_cx.select_all_or_error(&infcx);
+    if !errors.is_empty() {
+        return Err(errors);
+    } else {
+        return Ok(());
     }
 }
